requirements.txt
kafka-python==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
